requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
opencv-python
Pillow
tkinter
opencv-python
requests